requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
motor>=3.5.0
pymongo>=4.8.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.1
pydantic[email]>=2.8.0
python-multipart>=0.0.9
 